% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCor.R
\name{geneCor}
\alias{geneCor}
\title{Tool for Identification of CNAcor and METcor, Visualization of the distribution of expression of CNVcor genes and expression of METcor genes on a page, and Examination of the significance of each of those skewed distributions.}
\usage{
geneCor(
  cna = NULL,
  exp1 = NULL,
  alternative1 = c("two.sided", "less", "greater"),
  met = NULL,
  exp2 = NULL,
  alternative2 = c("two.sided", "less", "greater")
)
}
\arguments{
\item{cna, exp1, alternative1, met, exp2, alternative2}{}
}
\description{
Automatically individually compute the Pearson's correlation coefficients of genes shared between CNA data and EXP data, and those shared between MET data and EXP data, visualize the distribution of expression of CNVcor genes and expression of METcor genes on a page, and examine the significance of each of those skewed distributions. For further information on requirements as well as how to implement this tool, please visit my Github repository: https://github.com/huynguyen250896/geneCor
}
\examples{
geneCor(cna = df1, exp1 = df2, alternative1="less", met = df3, exp2 = df4, alternative2="greater")

}
